polynomialset: be cautious about label_t = letter_t
In some cases we can't decide a label_t is a letter_t, even in
letterized mode.  Consider for instance the case of a labelset which
is letterized, and that we then wrap in a nullableset: label_t is
pair<letter_t, bool>, which is not the same as bool.

* vcsn/core/rat/expressionset.hh: Not letterized.
* vcsn/weightset/polynomialset.hh (print_): Instead of checking lan and lan
(both being scheduled for removal), depend on whether is letterized.
* vcsn/labelset/oneset.hh (is_letterized): Of course not.