labelsets: concat -> mul
It clearly appears that there are many advantages of having a common
interface between labelsets and weightsets, typically to be able to
use weightset in place of labelsets for instance.

However, the product in labelsets is "concat", it is "mul" in
weightsets.  On occasions it results in really ugly code; for instance
in polynomials:

    add_here(res,
             labelset()->concat(lm.first, rm.first),
             weightset()->mul(lm.second, rm.second));

Besides, we already have ldiv and rdiv which are "reciprocals" to
"concat"...

Let's fix all this and rename "concat" as "mul".

This is not entirely trivial, in particular because "concat" could be
used with values of different nature: (letter, word) -> word etc., not
just (value, value) -> value.

* vcsn/alphabets/char.hh, vcsn/alphabets/setalpha.hh (concat): Rename as...
(mul): this.
* vcsn/labelset/genset-labelset.hh, vcsn/labelset/tupleset.hh: Likewise.

* tests/unit/label.cc, vcsn/algos/enumerate.hh, vcsn/algos/is-ambiguous.hh,
* vcsn/algos/is-partial-identity.hh, vcsn/algos/synchronizing-word.hh,
* vcsn/core/rat/expressionset.hxx, vcsn/weightset/polynomialset.hh:
Adjust.