dyn: generate the register signatures
Currently all the register declaration are performed with
REGISTER_DECLARE in each vcsn/algos/foo.hh.  First, this is tedious,
and second (actually that was my main motivation), it means that the
register implementations (lib/vcsn/algos/foo.cc) depend on their
headers, which is absolutely useless.

As a matter of fact, we should probably fuse all the dyn::foo
implementation in a single file: there is no real value to be gained
from having so many foo.cc files: lot of duplicate
instantiation/compilation from the compiler, and a lot of useless work
for the linker.

So as a first step, "parsing" the bridge declarations from the foo.hh
headers, to generate a single registers.hh file.

* build-aux/bin/registers-gen: New.
* vcsn/local.mk: Use it.

* vcsn/algos: Remove all uses of REGISTER_DECLARE.
* vcsn/dyn/fwd.hh (REGISTER_DECLARE): Remove.

* lib/vcsn/algos: Don't include the corresponding algo header, rather
include vcsn/dyn/registers.hh.

* vcsn/ctx/instantiate.hh: Include it too.