rmul: use it again in derivation
We've have the same bug uncovered twice.

The right multiplication of a polynomial of expressions by a weight
must not rmul the weights of the polynomials, but the labels (i.e.,
the expressions).  This was first reported by Sylvain Lombardy:

    commit 8830f2a72ca3a9cd7e239a021a0b6b297b888660
    Author: Akim Demaille <akim@lrde.epita.fr>
    Date:   Thu Jan 23 17:36:05 2014 +0100

    derivation: bug fix in right external product

    Right-multiply each monomial by the given weight instead of using the
    polynomial set external product.  Reported by Sylvain Lombardy.  Typed
    by Luca Saiu :-).

    * vcsn/algos/derivation.hh: Here.

Unfortunately the fix was not at the right place: instead of stopping
using polynomialset::rmul, the above fix open-coded a fixed version.
So another occurrence of the same bug was discovered later, in split:

    commit 48d911a0eeb0e4b40b44be50e7c9069bbe66d47b
    Author: Akim Demaille <akim@lrde.epita.fr>
    Date:   Tue Feb 17 15:00:23 2015 +0100

    polynomials: fix rmul

    * vcsn/weightset/polynomialset.hh (rmul): Fix.
    * tests/python/derivation.dir/ext-prod-breaking.gv: New.
    * tests/python/derivation.py: Check it.
    * tests/python/split.py: Enable a test that used to fail.

But then, the open-coded version in derivation should be
de-duplicated.

* vcsn/algos/derivation.hh (rweight): Use rmul.