minimize: weighted: really build the signatures, don't use transition_maps
The previous code was quite complex because it tries to avoid storing
the signatures: it just computes them on the fly when needed, i.e.,
when accessing the unordered map (the hash and equals_to functions).

Because hashing needs to be robust to the order of the transitions,
transition_maps were useful.

In the end, it is easier to genuinely build these signatures, and
leave the management of their lifetime to the map (where they are used
as keys).  Then, we don't even have to use transition_maps.

Compared to the previous version:

  1.55 1.35: a.minimize("weighted") # a = std([a-g]{300}), c = [a-k] -> B
  1.63 1.41: a.minimize("weighted") # a = std([a-g]{300}), c = [a-k] -> Z

* vcsn/algos/minimize-weighted.hh: here.