quotient: collapse partitions of partitions
When we run repeatedly cominimize and minimize, we get a
partition_automaton of a partition_automaton of a...  The usefulness
of such level of details is unclear.  What is clear though, is the
cost: it requires a lot of compilations for these stacked
partition_automata.

Let's avoid that.

Likewise, let cominimization return a partition_automata instead of a
transpose_automaton<partition_automaton<transpose_automaton>>.

* vcsn/core/partition-automaton.hh (make_partition_automaton): New.
* vcsn/algos/quotient.hh (quotient): Use it.
* vcsn/algos/minimize-moore.hh, vcsn/algos/minimize-signature.hh,
* vcsn/algos/minimize-weighted.hh, vcsn/algos/minimize.hh:
Use it to adjust the output type.

* doc/notebooks/automaton.cominimize.ipynb,
* doc/notebooks/automaton.costandard.ipynb,
* doc/notebooks/automaton.minimize.ipynb:
Update.