infiltration: fix a long standing bug
           .
         a/ \
          \ /  a
Let a = -> 0 ----> 1 ->

When we compute its self-infiltration, starting from state (0,0), we
first introduce the "product" transitions.  There are four, for the
there are destinations: (0,0), (0,1), (1,0), and (1,1).

Then when we introduce shuffle transitions, there are those were tape
0 only moves, arriving to (0,0) and (1,0), and then those where tape 1
only moves, arriving to (0,0) and (0,1).

So not only do we have three times (0,0) as destination, we also
arrive twice to (0,1) (and twice to (1,0)).  So we cannot use
new_transition in that case.

Of course, as a consequence, the performances of infiltration are
ruined.  If someone thinks this is really a problem, then she should
look for better optimizations:

  $ v score-compare --only 'product|infil|shu' v2.0-631-g1a020c3 <this commit>
  3.39  8.31  a.infiltration(a)    # a = std([a-e]?{30})
  4.05  3.87  a.product(a)         # a = std([a-e]?{50})
  0.80  0.82  a.product(a)         # a = thompson([a-e]?{80})
  3.21  3.05  a.shuffle(a)         # a = std([a-e]?{50})

* vcsn/algos/product.hh (add_shuffle_transitions_): Parameterize by a
Boolean to possibly force the use of add_transition instead of
new_transition in the case of infiltration.
* tests/python/infiltration.py: Check it.