infiltration: fix variadic calls
What we're doing is dead wrong.

Variadic infiltration is not trivial to implement, it's not just
product and shuffle in series.  For instance, consider three automata:

            <x>a
  x = -> 0 -----> 1 ->

and likewise for y and z.  Let's use # to denote infiltration.  In
(x # y) there is a transition ((0,0), <xy>a, (1,1)) coming from the
product-like transitions.

Therefore in (x # y) # z there is a transition
((0,0),0), <xy>a, ((1,1),0) by virtue of the shuffle-like transitions.

This kind of transition that mixes product and shuffle would never
appear in a naive implementation with only product and shuffle
transitions, but no combinations.

* vcsn/algos/product.hh: Fix by calling repeatedly the binary
infiltration.
* tests/python/infiltration.py: Check.