dyn: initiate changes to compile several algorithms at once
Each time we display an automaton in IPython, we compile dot (for SVG
rendering) _and_then_ info (for text rendering: IPython always calls
both to store both outputs in the notebook).

This is silly.

Fortunately dot and info have exactly the same signature, so as a
first experiment, it's quite easy.

* lib/vcsn/dyn/translate.cc: Be ready to compile groups of algorithms.
When dot is, also compile info.