automaton: introduce make_fresh_automaton
We often need to create an empty automaton from the context of another
one.  In that case, we might need to use "real_context", if the
automaton is a focus_automaton.  And we must think to use
automaton_nocv_t.

Let's make this simpler: make_fresh_automaton will address this task.
And to be more consistent, let's rename automaton_nocv_t as
fresh_automaton_t.

* vcsn/algos/copy.hh (make_fresh_automaton): New.
Use it where appropriate.

* vcsn/algos/are-equivalent.hh, vcsn/algos/concatenate.hh,
* vcsn/algos/determinize.hh, vcsn/algos/filter.hh,
* vcsn/algos/focus.hh, vcsn/algos/is-valid.hh,
* vcsn/algos/left-mult.hh, vcsn/algos/pair.hh, vcsn/algos/product.hh,
* vcsn/algos/quotient.hh, vcsn/algos/reduce.hh, vcsn/algos/scc.hh,
* vcsn/algos/star.hh, vcsn/algos/transpose.hh,
* vcsn/algos/universal.hh, vcsn/core/automaton-decorator.hh,
* vcsn/core/mutable-automaton.hh, vcsn/core/partition-automaton.hh,
* vcsn/core/permutation-automaton.hh, vcsn/ctx/traits.hh:
Adjust.