shortest: use boost::optional to mean "no argument"
To use "0" to mean "infinity" for both len and num is a problem: in
the case of len, it does make sense to ask for words of len = 0:
that's \e.  Currently there is no means to say that.

So instead of using 0 as default value, let's use boost::none as
default value to mean infinity.

* vcsn/algos/shortest.hh: Use boost::optional to denote infinity.
* vcsn/misc/name.hh: Add support for boost::optional<unsigned>.
* python/vcsn_cxx.cc, vcsn/ctx/instantiate.hh, vcsn/dyn/algos.hh:
Adjust.