context: accept escaped characters
In Python, when we run

   vcsn.context('lal_char(\x01-\xfe), b')

we actually get support from 1 to 195 (instead of xFE = 254).  Note
that the decoding is left to Python above: Vcsn does not receive the
escaped sequence, but its decoded value.

It turns out that Python does generate byte xFE, but the Boost.Python
conversion delivers a UTF-8 conversion of it, which is a sequence that
starts with 195.

I think Vcsn should not try to understand characters at all, it should
really receive raw strings.  But Python wants to understand these
strings as Unicode, otherwise it won't be able to display these
strings!

So in the end, I think the safest is _not_ to try to fix the Python
line above, but rather to support escaped character in the definition
of the context:

   vcsn.context('lal_char(\\x01-\\xfe), b')

or

   vcsn.context(r'lal_char(\x01-\xfe), b')

* vcsn/alphabets/char.hh (get_char): Support hexadecimal escapes.
* vcsn/alphabets/setalpha.hh (get_word): Be ready for \-escapes too.
* tests/unit/label.cc: Check it.
Also, check that we do support all these letters.
* tests/python/context.py: Check it from Python too.