expressionset: simplify
* vcsn/core/rat/expressionset.hxx (is_unweighted_nonsum_, is_nonsum_)
(mul_products_, mul_unweighted_nontrivial_products_): Remove.
These functions are wrong on several accounts:
- don't name a function 'nonstuff', just return 'stuff', and use !
- don't use switch instead of an if
- use type_ignoring_lweight_ where it make sense.

In the end, one can see that the condition that used this function is
a straightforward consequence of the previous tests, so the dead code
can be removed...
And as a matter of fact, the dead code and the function that hid it
are equivalent, so just keep the former, as it is clearer.