letterize: style: split letterized_labelset implementation
letterized_labelset is a trait, very similar to the already existing
nullableset_traits and law_traits.  However the former is implemented
in a single place, letterize.hh, while the latter two are implemented
in the corresponding labelset header.  The second scheme seems more
extensible, so dispatch the implementation of letterized_labelset
where appropriate.

* vcsn/algos/letterize.hh (letterized_labelset): Dispatch to...
* vcsn/labelset/labelset.hh, vcsn/labelset/letterset.hh,
* vcsn/labelset/nullableset.hh, vcsn/labelset/tupleset.hh,
* vcsn/labelset/wordset.hh: here.