fresh_automaton: prepare for a possible change of context
Sometimes we wish to create a fresh automaton, but with a different
context.  For instance, proper needs to create a fresh automaton,
however, it would be a pity to preserve a lan labelset, we really want
to change its context.

* vcsn/algos/determinize.hh, vcsn/algos/focus.hh, vcsn/algos/synchronize.hh,
* vcsn/algos/transpose.hh, vcsn/core/automaton-decorator.hh,
* vcsn/core/mutable-automaton.hh, vcsn/core/partition-automaton.hh,
* vcsn/core/permutation-automaton.hh
(fresh_automaton_t): Template it with the wanted context.
* vcsn/ctx/traits.hh (fresh_automaton_t_of): Accept this optional second
template parameter.