labelset: a version of convs that takes a lambda
The current interface, convs(istream) -> set<letters_t>, is
inconvenient in several regards.  It is costly (building sets on
several layers), and hard to extend to support negated classes on the
caller site, since that requires to use the alphabet, which is a
problem with expressionset, oneset, tupleset, etc.

Actually, everything becomes simpler if the negated classes are
handled at the bottom layer.  Better yet: don't return a container,
but apply a lambda.  Then the innermost implementation can even use
itself recursively in the negated case to get the list of letters to
remove from the alphabet.

For a while I have considered stopping using a std::set of letters at
the innermost layer, but it has many consequence.  For instance [ab]
and [ba] are no longer equivalent, and [a] is also different from [aa]
(which denotes <2>a).  The first issue, [ab] != [ba] could be accepted
on account that a+b is already different from b+a.  However, [aa] == <2>a
is really ugly: these are no longer letter _classes_, since the number
counts.

If in the future we support classes on automata, we will probably not
ensure the user original order, but rather focus on something
computationally efficient: sorted classes.

So let's keep sorted classes for the moment, even if that means an
additional cost: the set that ensures the unicity.  If we really want
to avoid this cost, we have the option of adding say a template
parameter to conv to decide whether the result should be
sorted+uniques or not.

* vcsn/labelset/genset-labelset.hh (convs_): Rewrite to apply a functor
instead of generating a sets.

* vcsn/core/rat/expressionset.hh, vcsn/labelset/letterset.hh,
* vcsn/labelset/nullableset.hh, vcsn/labelset/oneset.hh,
* vcsn/labelset/tupleset.hh, vcsn/labelset/wordset.hh,
* vcsn/weightset/polynomialset.hh:
Adjust.