proper: add a proper_separate algorithm
Perform the proper algorithm in two separate automata, a lal and a lao.

Proper is now in its own file, separated from the Epsilon Remover algorithm,
which can be present in multiple versions (for now, epsilon-remover and
epsilon-remover-separate).
Separate other unrelated tool functions to factor the code (debug_level and
derived contexts).

* vcsn/algos/proper-separate.hh: Create.
* vcsn/misc/star_status.hh: Move the start status from proper.hh
* vcsn/algos/proper-separate.hh -> epsilon-remover-separate.hh: Rename.
* vcsn/algos/is-valid.hh: Use epsilon_remover instead of in_situ_remover.
* vcsn/algos/proper.hh: Move the algorithm to vcsn/algos/epsilon-remover.hh.
* vcsn/misc/debug_level.hh: Move debug_level() from proper.hh.
* vcsn/misc/derived_context.hh: Move {proper,ones}_context from proper.hh.
* vcsn/local.mk: Add the created files.