weight_series: new algorithm
This algorithm computes the sum of the weights of the language. For most
weightsets, if there is a non-zero cycle the algorithm doesn't work. For
tropical semi-rings, if there is a negative (resp. positive) cycle for min
(resp. max), the algorithm doesn't work.

The ideal use case is for <N, +, min>, and the algorithm has been specialized
for this case (it comes down to a shortest path).

* vcsn/algos/weight-series.hh: new.
* python/vcsn_cxx.cc, vcsn/dyn/algos.hh, vcsn/local.mk: Register.