rat: introduce the "traditional" level of identities
Expression support a means to specify the identities they use.
Currently there is "trivial" (which should actually be named
"associative" as both sum and product are actually associative), and
"series".  Let's introduce "traditional" which is in between: compared
to "trivial" it adds commutativity and "idempotence" of sum (<k>a+<h>a
=> <k+h>a), compared to "series" it does not force distributivity.

* vcsn/core/rat/identities.hh, lib/vcsn/rat/identities.cc (traditional): New.
* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx:
Use it.
* tests/rat/br.rat, tests/rat/z.rat: Check it.