rat: expressionset: provide a default set of identities
Currently the ctor of expressionset and make_expressionset require an
explicit definition of the identities.  This not so nice, as the
default value (trivial currently) is spread in many different places.
This is risky, as is any kind of duplication.

Rather, let's provide them (expressionset's ctor and
make_expressionset) with a default level of identities.

While at it, let's make identities a real class, so that we can
benefit from having a ctor to define the default identities set.

* vcsn/core/rat/identities.hh, lib/vcsn/rat/identities.cc (identities):
Turn into a real class.

* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx: Use it.

Adjust dependencies.