rat: introduce the "trivial" level of identities
Well, so far what we called "trivial" is actually enabling
associativity of sum and product, which goes against the desires of
Jacques and Sylvain.  Now that we support several levels of
identities, let's rename the former "trivial" into "associative"
(which "becomes" the default), and introduce a lower level under the
name "trivial".

* vcsn/core/rat/identities.hh, lib/vcsn/rat/identities.cc (associative):
New.
This is the new default.
* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx: Adjust.
* doc/notebooks/Expressions.ipynb: Document that.