expressionset: linear: reimplement insertion in a sum
Instead of copying the sum, and then to insert the new element, copy
the lower part, then the element, then the upper part.  Of course, pay
attention to the case where the element is already part of the sum.

Unfortunately, there is no visible improvements (nor regression).  But
it's more natural this way, and a bit shorter (30 insertions, 42 deletions).

While at it, we actually fix a nasty bug: a+a (linear) would actually
result in a sum of a single element, <2>a.

* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx
(remove_from_sum_series_): Remove.
(insert_in_sum_series_): Do that.