expression: fix a parsing issue
Currently <2>a<3>b is parsed as (<2>a)(<3>b), which right, but
<2>a<3>(b) results in <6>(ab), which is wrong.  This is because in the
parser, the precedence of "(" is lower than that of "weight", so at
this point:

   <2>a    <3>   .  (
   exp  "weight" . "("

the reduction is performed first, leading to "<6>a" and then the
parsing proceeds.  So obviously the precedence should be higher.

* lib/vcsn/rat/parse.yy: Fix predences.

* tests/rat/br.rat: ab(<2>\e) was expected to be parsed as ab<2>,
it should have been a<2>b.