efsm: support I/O of weights
It turns out that fstinfo does report the type of the weightset, as
"arc type".  AFAICT, there are three pre-defined such arc types:
"standard" (which corresponds to our rmin), "log" and "log64" (which
correspond to our log, which is on double, like log64).

It turns out that our only check that uses I/O of weights was on
binary.gv, which is a Z-automaton, which was interpreted as a
Zmin-automaton by OpenFST, and (bad luck), its squarre is the same in
both weightsets!  (Because 2 + 2 = 2 * 2).

So let's use better Zmin automata.

* bin/efstdecompile: Extract and report the arc_type.

* vcsn/algos/edit-automaton.hh, lib/vcsn/algos/edit-automaton.cc
(weightset_type, weightset_type_, weightset): New.
Use them to compute the weightset name.

* lib/vcsn/algos/efsm.cc: By default, use "standard" as arc type.

* vcsn/algos/efsm.hh: Report the arc_type.

* python/vcsn/automaton.py (fstcat): New.

* tests/python/efsm.dir/a1.efsm,
* tests/python/efsm.dir/a2x.efsm,
* tests/python/efsm.dir/a2xyz.efsm,
* tests/python/efsm.dir/abs.efsm,
* tests/python/efsm.dir/one.efsm,
* tests/python/efsm.dir/str.efsm:
Set the arc type.
* tests/python/efsm.py: Adjust.

* tests/python/efsm.dir/lal-char-z.efsm: Rename as...
* tests/python/efsm.dir/lal-char-zmin.efsm: this.
* tests/python/efsm.dir/lal-char-z.gv: Rename as...
* tests/python/efsm.dir/lal-char-zmin.gv: this.
* tests/python/efsm.dir/lat-z.efsm: Rename as...
* tests/python/efsm.dir/lat-zmin.efsm: this.
* tests/python/efsm.dir/lat-z.gv: Rename as...
* tests/python/efsm.dir/lat-zmin.gv: this.
* tests/python/efsm.dir/binary.efsm: Rename as...
* tests/python/efsm.dir/slowgrow.efsm: this.