identities: linear: E<k> => <k>E
Currently we have something which is bizarre: the "distributive"
identities require a commutative semiring and implement the following
laws:

  <k>E<h>F => <kh>(EF)     (1)
  <k>(E+F) => <k>E + <k>F  (2)
  (E+F)<k> => E<k> + F<k>  (3)

However a*<2> is in normal form.

I think this is wrong.  What is actually missing is

  E<k> => <k>E             (4)

in which case (2) suffices, we don't need (3).  Jacques Sakarovitch
seems to agree with that (there are no left notes that I could exploit
from his conversations with Luca Saiu, who initiated this
"distributive" identities).

Actually, (2) (and (3)) are clearly disbributive laws, but (1) and (4)
don't fit in this category.  And they seem very natural.  So I hereby
move them from "disbributive" to "linear".

Alternatively, we could have introduced another level of identities,
bw linear and disbributive.  Also, instead of _requiring_ the weights
to be commutative which building the expressionset, I prefer to enable
these identities only when the weightset is commutative.

* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx
(mul_, mul_expressions_, mul_series_, nontrivial_mul_expressions_)
(nontrivial_mul_series_): Remove.
That's too hairy.
(mul): Handle all these subcases.
(rmul): Map to lmul when commutative and linear.
(power): Adjust.
* doc/notebooks/Expressions.ipynb: Document.

* tests/bin/rat.py, tests/python/expression.py,
* tests/python/infiltration.py, tests/python/info.py,
* tests/python/kmul.py, tests/python/lift.py,
* tests/python/multiply.py, tests/python/reduce.py,
* tests/python/shuffle.py, tests/python/standard.py,
* tests/python/thompson.py, tests/python/to-expression.py,
* tests/python/weight.py, tests/rat/b.rat,
* tests/rat/common-weights.rat, tests/rat/common.rat, tests/rat/z.rat:
Adjust the test cases.
In some cases, I preferred to move the test to the 'associative' level
of identities, because it's better suited to exercise some parts of
the code.
At other places, let's just adjust our expectations.