cominimize: fix it
Shame on me: cominimize worked only on... transposed automaton, which
quite defeats the whole point.  And the test did not catch this,
because, to avoid having to deal with too many reference files, we
applied the cominimization on the transposed input, so that we could
expect the transpose result as result...

Having more expected result for cominimization is indeed not nice.  So
instead of working an a transposed automaton (i.e., a
transpose_automaton decorator), copy the transposed automaton as a
forward automaton.

* vcsn/core/partition-automaton.hh (quotient_t_of): New.
* vcsn/algos/quotient.hh: Use it.
* tests/python/minimize.py: Check cominimization.