rat: use dyn::expression and remove dyn::expressionset from the parser
Almost since the beginning Vcsn used dyn::expressionset to parse the
expressions.  Expressions are stored as rat::exp_t, the (non
templated) base class for all the expression nodes.

This commit moves the parser to rather use dyn::expression to store
the expressions in the parser, using dyn:: in the actions.

We do have a 2x performance penalty (due to moving from virtual-based
dispatch to dyn::-dispatch):

 0.57  1.09 b.expression(e)      # e = "(\e+a)" * 500, 100x

but I think I can live with that.  In particular, I expect the parser
to be able to do more things on top of dyn::expression (which allows
to easily mix different expressionsets) rather than being bound to a
specific dyn::expressionset.

* lib/vcsn/rat/parse.yy: Move to parsing with dyn::expression.
* lib/vcsn/rat/driver.hh, lib/vcsn/rat/driver.cc: Adjust.