dyn: get rid of dyn::expressionset
Almost from the start of Vcsn we introduced dyn::expressionset (under
a different name: it took quite a while until things got cleaned,
understood, classified, renamed, moved, etc.).

Its performs type-erasure on an expressionset, and its main purpose
was the parser which used it as an expressionset, for instance to run
`rs.mul($1, $2)` and the like.  This required that $1, $2 etc. were
type-erased expressions.  It would seem normal that they were
dyn::expression, but they were not: they were rat::exp_t, a base class
for all the expression AST nodes.

However, dyn::expressionset was sort of an anomally, since it is a
ValueSet, although in dyn:: there should be only dyn:: values which
bind together a Value and a ValueSet.  The proper design was rather to
use dyn:: routines in the parser, hence to use dyn::expressions
instead of rat::exp_t.  So we should also be able to get rid of it.

* vcsn/dyn/expressionset.hh, vcsn/dyn/expressionset.hxx: Remove.

* lib/vcsn/algos/read.cc (read_expression): Use ctx+ids instead of
expressionset.
* lib/vcsn/rat/driver.hh, lib/vcsn/rat/driver.cc: Likewise.
* lib/vcsn/rat/read.hh, lib/vcsn/rat/read.cc: Ditto.
* vcsn/algos/copy.hh: And again.

* lib/vcsn/rat/parse.yy: Adjust.
* vcsn/core/rat/expressionset.hxx: Adjust.

* vcsn/algos/make-context.hh (context_of_expressionset)
(make_expressionset): Remove.