rat: use dyn::weight to store weights
So far we used a trick to store weights: they were stored as <k>\e
(similarly to labels being stored as expressions), enjoying the fact
that expressions were stored as rat::exp_t, the non-template base
class of all the expression.  The parser was recently converted to use
dyn::expression, and now, let's use dyn::weight for weights.

* lib/vcsn/rat/parse.yy: here.