focus: move to a better place
The routines to focus on a given tape (for a context, an
expressionset, etc.) are used in many places (e.g., in polynomialset).
Currently they are defined in vcsn/algos, which is wrong: algos can
depend on vcsn/labelset etc. but not the converse.

So move focus into a lower layer.

* vcsn/algos/focus.hh (focus_labelset, focus_context): Move to...
* vcsn/labelset/tupleset.hh: here.