rat: tuple: use the traditional a|x notation instead of (a,x)
When printing a tuple in the case of a label, print "a|x".  In the
case of weights, print "(1,1/2)".  Eventually, we should probably also
get rid of the parens in this case too.

* vcsn/misc/format.hh (for_labels, for_weights, is_for_labels)
(is_for_weights): New.
* vcsn/weightset/polynomialset.hh (print_weight_, print): Use it.
* vcsn/labelset/tupleset.hh (print): Use it.

But then, in order to be able to read the automata in dot format, we
also need to support a|x.

* vcsn/labelset/tupleset.hh (eat_separator_): Accept either ',' or '|'.

* tests/bin/update-test, tests/python/compose.dir/result.gv,
* tests/python/compose.py, tests/python/lift.py,
* tests/python/polynomial.py, tests/python/reduce.py,
* tests/python/synchronize.dir/bool.gv,
* tests/python/synchronize.dir/bool_delay.gv,
* tests/python/synchronize.dir/bool_min.gv,
* tests/python/synchronize.dir/bool_min_delay.gv,
* tests/python/synchronize.dir/z.gv,
* tests/python/synchronize.dir/z_delay.gv,
* tests/python/synchronize.dir/z_min.gv,
* tests/python/synchronize.dir/z_min_delay.gv,
* tests/python/transducer.py:
Adjust.