rat: tuple: support variadic tuples
Currently we parse "|" as a binary operator, so a|b|c is actually
parsed as (a|b)|c, a 'lat<lat<lan, lan>, lan>' instead of a 'lat<lan,
lan, lan>'.

Besides, because of this we build too many contexts (cause a serious
slowdown): 'lan' for a, b and c, 'lat<lan, lan>' for a|b, and finally
'lat<lat<lan, lan>, lan>'.  Only 'lan' and 'lat<lan, lan, lan>' are
needed.

So stop treating "|" as a regular operator: let it actually stack
several expressions, that are tupled together on completion.

* lib/vcsn/rat/parse.yy (sum, tuple.1, tuple): New non-terminals, used
to handle precedence and associativity by hand, to allow tuple.1 to
use a stack of expressions.
* tests/rat/tuple.rat: New.