rat: tuple: let "a|x" denote a multitape label, not a tuple of expressions
* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx (tuple):
Add a new rewriting rule that rewrite tuples of expressions that are labels
as a multitape-label.
* tests/python/transducer.py: Exercise this by calling standard on such
an expression.

This revealed a flaw in the pretty-printer: it thinks that "a|x" is
self-delimited and can be printed with parens.  It used to be true,
since we used to print it as "(a, x)".  Now, it's no longer the case,
so we pretty-print "(a|x)*" as "a|x*".

* vcsn/core/rat/printer.hh (is_word_): Fix the support for tuples.