lift: style: use enable_if to select the implementation
The traits lift relies on has two implementation: one when no tape is
given (actually meaning all of them), and another when a tape is
given.  But since parameter packs accepts no parameters at all, to
tell the difference bw the two, the second one has two parameters: the
FirstTape and the other Tapes.

Use enable_if instead to reject when Tapes is empty.

* vcsn/algos/lift.hh (lifted_context_tape_t): Rework as...
(lifted_context_tape_impl): this.