lift: style: avoid split implementation
Currently the case of full-lifting is implemented as a "free" set of
traits, and the case of per-tapes lifting is handled by a
"centralized" traits: lifted_context_tape_impl.  But then, in the case
of full-lifting, it bounces to the free traits.

This is useless complex.

* vcsn/algos/lift.hh: Move all the free-traits handling the
full-lifting case into the full-lifting specialization of
lifted_context_tape_impl.