dyn: fix the parsing of label classes
It turns out that we are not able to parse types such as const
std::set<std::pair<std::string, std::string>>, which we use for label
classes (a set of label ranges).

We never saw it, because no one used label ranges with a non-prebuilt
context.  Since I started to remove some of these prebuilt contexts, I
just saw the failure.

* vcsn/dyn/context-parser.hh, lib/vcsn/dyn/context-parser.cc
(parameters_): New.
(any_): Use it for all the templates, instead of special casing them
all.