tuple: fix transposition of words
Running trie on a LAW x LAW labelset breaks since the introduction of
cotrie.  That's because its call to transpose is invalid.

* vcsn/labelset/tupleset.hh, vcsn/labelset/nullableset.hh (transpose):
Fix to work on words.

This fix for nullableset is not robust enough: when nullable wraps a
labelset that does not support one, it works on the labels of the
inner labelset.  But it's certainly not ready to work on _words_.

Since I'm unsure of the future of nullableset, let's not waste time on
this now.