rat: optimize `less`
Do not recurse into two expressions that are actually physically the
same (same address).

This is particularly beneficial to derived-term computations (both
expansions and derivations), as derived-terms have often common
subexpressions that are actually equal (shallow copies, not deep).

On erebus, 36s, on c = [a-z] -> Z, and r = (a+b)*b(<2>a+<2>b){N}

 0.33  0.13 40% r.derived_term("derivation") # N = 150
 0.32  0.11 34% r.derived_term("derivation") # N = 150
 0.24  0.03 13% r.derived_term("expansion")  # N = 150
 1.47  0.68 46% r.derived_term("derivation") # N = 300
 1.40  0.59 42% r.derived_term("derivation") # N = 300
 9.23  1.12 12% r.derived_term("expansion")  # N = 300, 10x

It's even more beneficial to expansions: almost 9x speed up on these
benches.

* vcsn/core/rat/less.hh: Do that.