to-expansion: optimize the termination case
Currently to-expansion and derivation have similar ideas to handle
variadic multiplication, but different implementations.

In derivation, when we process an operand, we immediately multiply its
derivative by all the remaining right-hand side argument.

In expansion, when we process an operand, we first multiply the
previous result with it.

That's equivalent, but it turns out that the former is much faster
than the latter.  See the last line below.

 0.13  0.13 dt("derivation") # r = (a+b)*b(<2>a+<2>b){150}, c = [a-z] -> Z
 0.11  0.11 dt("derivation") # r = (a+b)*b(<2>a+<2>b){150}, c = [ab] -> Z
 0.03  0.03 dt("expansion")  # r = (a+b)*b(<2>a+<2>b){150}, c = [a-z] -> Z

 0.66  0.69 dt("derivation") # r = (a+b)*b(<2>a+<2>b){300}, c = [a-z] -> Z
 0.59  0.60 dt("derivation") # r = (a+b)*b(<2>a+<2>b){300}, c = [ab] -> Z
 1.12  1.13 dt("expansion")  # r = (a+b)*b(<2>a+<2>b){300}, c = [a-z] -> Z, 10x

 0.78  0.79 dt("derivation") # r = a?{100}, c = [a] -> Z
 9.41  0.50 dt("expansion")  # r = a?{100}, c = [a] -> Z

* vcsn/algos/to-expansion.hh (prod): Stop iterating continuously on
the current result, rather work on what is about to be added.
This allows to simplify the termination case.