derivation: simplify
* vcsn/algos/derivation.hh (prod): The code pretends that we finish our
iterations by multiply the result with all the remaining operands.
However we already took care of that in the main loop.
So it should actually break the result.
It turns out that it does not, because the computations are not performed
on what is going to be returned!!!
So, that's just removing dead code.