rat: none: improve
The introduction of the variadic perfect forward ctor for variadic is
beneficial for a few scores:

 0.78  0.58 a.expression("associative", "naive") # a = std([a-d]?{100})
 1.16  0.86 b.expression(e)      # e = "(\e+a)" * 500, 100x
 0.77  0.40 b.expression(e)      # e = [ab]{20000}, 1000x

* vcsn/core/rat/expression.hh (variadic): Provide a variadic ctos.
* vcsn/core/rat/expressionset.hxx: Use it.
Clean up the handling of none: instead of checking whether identities
are enabled to apply then, just first handle the case where we don't
apply any identities.
Use consistent comments.
Use a consistent scheme to return the values.