rat: conv: support a format argument
In order to support multi-char letters, we will need to be able to
tell whether a stream contains a single "letter" (i.e., read till
EOF), or not (in which case, read the next char, or, if there are
quotes, read up to the closing quote).

* vcsn/labelset/letterset.hh, vcsn/labelset/nullableset.hh,
* vcsn/labelset/tupleset.hh, vcsn/labelset/wordset.hh
(conv): Accept a "quoted" argument.

* lib/vcsn/algos/read.cc, vcsn/algos/read.hh,
* vcsn/ctx/instantiate.hh, vcsn/dyn/algos.hh (read_label):
Ditto.

* lib/vcsn/rat/driver.cc: When read a letter, set quoted to false.