dyn: let contexts display "context" in their names
So far, contexts used to display themselves as
"letterset<char_letters>, b" for instance.  This is a problem for
instance if we have tuples that can contain labelsets, weightsets and
contexts: how do we know what tuple<lal_char, q> means?

Now, let's have the contexts declare themselves as "context<LS, WS>".
Fortunately, we can preserve the interface with the user.

* vcsn/ctx/context.hh (sname, make): Use the "context<LS, WS>" syntax.
* lib/vcsn/dyn/signature-printer.cc: Ditto.
* vcsn/dyn/context-parser.hh, lib/vcsn/dyn/context-parser.cc (context_):
Rework to accept the possible "context<...>" part.