dyn: get rid of call_variadic, use only tuples
See a couple of commits before: when a dyn:: routine receives a vector
of dyn::automata or dyn::expressions, they now forward them with a
tuple type (of each precise type) instead of a variadic template list.

This is much better, as, for instance, it allows to have other arguments,
such as the `lazy` argument of conjunctions.

So get rid of the previous system.

* lib/vcsn/algos/registry.hh (call_variadic): Remove.
* build-aux/bin/registers-gen: The call function is now always `call`.

* vcsn/misc/name.hh: Support vectors of dyn::expressions.

* vcsn/algos/conjunction.hh, vcsn/algos/tuple.hh: No longer use
variadic templates, but rather tuples.
* vcsn/ctx/instantiate.hh: Adjust.

* lib/vcsn/algos/focus.cc (conjunction): Remove, it is now generated
like the others.