mutable_automaton: minor optimizations
The following "optimization" are natural, yet they appear to have no
visible impact on score.  I will keep them as the code, albeit
slightly more complex, is more alike what one would do to save cycles.

* vcsn/core/transition.hh: Provide ctors so that we can use emplace_back
on containers of these guys.
* vcsn/core/mutable-automaton.hh: Use emplace_back instead of resize.
First to get the benefit of growth of emplace_back, but also to allow
direct construction instead of construction-and-then-assignment.