automaton.type: new
When displaying an automaton, IPython also registers the text
representation, which currently relies on info().  Some measures from
info evaluate the automaton via calls to out() or all_out().  This is
not nice: each time we display a lazy automaton, it grows.

The algorithms called by info() should probably not be tailored to
avoid evaluating lazy states, so, instead, provide the text
representation of automata without using info().

* vcsn/dyn/algos.hh, lib/vcsn/algos/others.cc (type): New.
* python/vcsn_cxx.cc: Add it.
* python/vcsn/automaton.py: Remove it.