to-expression: simplify
This implementation, based on a heap, is based on that of proper.  But
the latter concerns only states that have incoming spontaneous
transitions, and not all have.  Here, all the states are to be
eliminated, so there is no precautions to take about the presence of
the states in the heap: they are all their.

 0.12  0.12  0.11  0.12  0.12  0.12  0.11 a.expr("associative", "delgado") # a = ladybird(2000)
 0.08  0.08  0.08  0.08  0.08  0.09  0.08 a.expr("associative", "naive")   # a = ladybird(8000)
 0.55  0.54  0.55  0.56  0.55  0.62  0.56 a.expr("associative", "naive")   # a = std([a-d]?{100})
 0.19  0.19  0.19  0.20  0.19  0.21  0.19 a.expr("linear", "delgado")      # a = ladybird(2000)
 0.25  0.25  0.21  0.21  0.21  0.22  0.21 a.expr("linear", "delgado")      # a = std([a-d]?{15})
 0.38  0.38  0.38  0.38  0.38  0.42  0.38 a.expr("linear", "naive")        # a = ladybird(4000)
 0.14  0.14  0.14  0.14  0.14  0.15  0.14 a.expr("linear", "naive")        # a = std([a-d]?{9})
             ^^^^
             this commit

* vcsn/algos/to-expression.hh (profile): Remove.
Simplify dependencies.