tupleset: enforce shortlex ordering
Currently we have \e|a < \e|aa < \e|aaa ... < a|\e, i.e., the
lexicographical ordering.  We want the shortlex one.

This change highlights the fact that trie relies on `less`, which is
shortlex.  In the context of trie, this results in a less natural
automaton, and actually, a less efficient one if we relied on the fact
that transitions are sorted on labels.

But this is a different issue, so let's just fix the expected result
in trie now that tuple is fixed.  We will have to address this
elsewhere.

* vcsn/labelset/tupleset.hh (less): Here.
(size_): Be robust to valuesets that don't provide a size() (weightsets
typically don't).
* vcsn/core/rat/expressionset.hh, vcsn/core/rat/expressionset.hxx
(size): New.
* tests/python/trie.py: Adjust.